requirements.txt
Django
Pillow
MySQL-python
django-crispy-forms
django-contact-form